Corrigindo desafios streams

+++ src/streams/Streams_19_desafios/Desafio_12.java
@@ -14,7 +14,7 @@ Com a ajuda da Stream API, encontre o produto de todos os números da lista e ex
     public static void main(String[] args) {
         List<Integer> numeros = Arrays.asList(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 5, 4, 3);
 
-        int produto = numeros.stream()
+        long produto = numeros.stream()
                 .reduce(1,(a,b) -> a * b);
 
         System.out.println(produto);
